+++ controllers/libs/cert/certificates.go
@@ -1,132 +1,116 @@
 // Copyright (C) 2023 Red Hat
 // SPDX-License-Identifier: Apache-2.0
 
-// Package cert provides various utility functions regarding handling cert-manager for the sf-operator
+// Package cert provides various utility functions regarding handling the local CA for the sf-operator
 package cert
 
 import (
+	"bytes"
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
+	"math/big"
 	"time"
 
-	cmacme "github.com/cert-manager/cert-manager/pkg/apis/acme/v1"
-	certv1 "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
-	cmmeta "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"github.com/softwarefactory-project/sf-operator/controllers/libs/utils"
 )
 
 const (
 	LocalCACertSecretName = "ca-cert"
+	// Valid for 30 years, which should be way more than the expected runtime of a deployment.
+	validity = 30
 )
 
-var EonDuration, _ = time.ParseDuration("262800h") // 30 years
+func X509CA() (*x509.Certificate, *rsa.PrivateKey, *bytes.Buffer, *bytes.Buffer) {
+	caCert := &x509.Certificate{
+		SerialNumber: big.NewInt(1),
+		Subject: pkix.Name{
+			CommonName:   "SF Local CA",
+			Organization: []string{"Red Hat, INC."},
+		},
+		NotBefore:             time.Now().AddDate(-1, 0, 0),
+		NotAfter:              time.Now().AddDate(validity, 0, 0),
+		IsCA:                  true,
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
+		KeyUsage:              x509.KeyUsageCertSign | x509.KeyUsageCRLSign,
+		BasicConstraintsValid: true,
+	}
 
-func MkBaseCertificate(name string, ns string, issuerName string,
-	dnsNames []string, secretName string, isCA bool, duration time.Duration,
-	usages []certv1.KeyUsage, commonName *string,
-	privateKey *certv1.CertificatePrivateKey) certv1.Certificate {
-	renewBefore, _ := time.ParseDuration("168h") // 7 days
-	cert := certv1.Certificate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: ns,
-		},
-		Spec: certv1.CertificateSpec{
-			DNSNames:    dnsNames,
-			Duration:    &metav1.Duration{Duration: EonDuration},
-			RenewBefore: &metav1.Duration{Duration: renewBefore},
-			SecretName:  secretName,
-			IssuerRef: cmmeta.ObjectReference{
-				Kind: "Issuer",
-				Name: issuerName,
-			},
-			IsCA:   isCA,
-			Usages: usages,
-		},
+	caPrivKey, err := rsa.GenerateKey(rand.Reader, 4096)
+	if err != nil {
+		utils.LogE(err, "Unable to generate a private key for the local CA")
 	}
-	if commonName != nil {
-		cert.Spec.CommonName = *commonName
+
+	caBytes, err := x509.CreateCertificate(rand.Reader, caCert, caCert, &caPrivKey.PublicKey, caPrivKey)
+	if err != nil {
+		utils.LogE(err, "Unable to create local CA certificate")
 	}
-	if privateKey != nil {
-		cert.Spec.PrivateKey = privateKey
-	}
-	return cert
+
+	caPEM := new(bytes.Buffer)
+	pem.Encode(caPEM, &pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: caBytes,
+	})
+
+	caPrivKeyPEM := new(bytes.Buffer)
+	pkcs8, _ := x509.MarshalPKCS8PrivateKey(caPrivKey)
+	pem.Encode(caPrivKeyPEM, &pem.Block{
+		Type:  "RSA PRIVATE KEY",
+		Bytes: pkcs8,
+	})
+	return caCert, caPrivKey, caPEM, caPrivKeyPEM
 }
 
-func MkCertificate(name string, ns string, issuerName string,
-	dnsNames []string, secretName string, privateKey *certv1.CertificatePrivateKey, duration time.Duration) certv1.Certificate {
-	usages := []certv1.KeyUsage{
-		certv1.UsageServerAuth,
-		certv1.UsageClientAuth,
-		certv1.UsageDigitalSignature,
-		certv1.UsageKeyEncipherment,
+func generateSerialNumber() (*big.Int, error) {
+	serialNumber, err := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), 128))
+	if err != nil {
+		return nil, err
 	}
-	return MkBaseCertificate(
-		name, ns, issuerName, dnsNames, secretName, false, duration, usages, nil, privateKey)
+	return serialNumber, nil
 }
 
-func IsCertificateReady(cert *certv1.Certificate) bool {
-	for _, condition := range cert.Status.Conditions {
-		if condition.Type == "Ready" && condition.Status == "True" {
-			return true
-		}
+func X509Cert(caCert *x509.Certificate, caPrivKey *rsa.PrivateKey, dnsNames []string) (*bytes.Buffer, *bytes.Buffer) {
+	serialNumber, err := generateSerialNumber()
+	if err != nil {
+		utils.LogE(err, "Unable to generate a serial number for X509 certificate")
 	}
-	return false
-}
 
-func MkLetsEncryptIssuer(name string, ns string, server string) certv1.Issuer {
-	return certv1.Issuer{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: ns,
+	cert := &x509.Certificate{
+		SerialNumber: serialNumber,
+		DNSNames:     dnsNames,
+		Subject: pkix.Name{
+			CommonName:   "Certificate",
+			Organization: []string{"Red Hat, INC."},
 		},
-		Spec: certv1.IssuerSpec{
-			IssuerConfig: certv1.IssuerConfig{
-				ACME: &cmacme.ACMEIssuer{
-					Server:         server,
-					PreferredChain: "ISRG Root X1",
-					PrivateKey: cmmeta.SecretKeySelector{
-						LocalObjectReference: cmmeta.LocalObjectReference{
-							Name: name,
-						},
-					},
-					Solvers: []cmacme.ACMEChallengeSolver{
-						{
-							HTTP01: &cmacme.ACMEChallengeSolverHTTP01{
-								Ingress: &cmacme.ACMEChallengeSolverHTTP01Ingress{},
-							},
-						},
-					},
-				},
-			},
-		},
+		NotBefore:   time.Now().AddDate(-1, 0, 0),
+		NotAfter:    time.Now().AddDate(validity, 0, 0),
+		ExtKeyUsage: []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
+		KeyUsage:    x509.KeyUsageDigitalSignature,
 	}
-}
 
-func MkSelfSignedIssuer(name string, ns string) certv1.Issuer {
-	return certv1.Issuer{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: ns,
-		},
-		Spec: certv1.IssuerSpec{
-			IssuerConfig: certv1.IssuerConfig{
-				SelfSigned: &certv1.SelfSignedIssuer{},
-			},
-		},
+	certPrivKey, err := rsa.GenerateKey(rand.Reader, 4096)
+	if err != nil {
+		utils.LogE(err, "Unable to generate a private key for X509 certificate")
 	}
-}
+	certBytes, err := x509.CreateCertificate(rand.Reader, cert, caCert, &certPrivKey.PublicKey, caPrivKey)
+	if err != nil {
+		utils.LogE(err, "Unable to create X509 certificate")
+	}
 
-func MkCAIssuer(name string, ns string) certv1.Issuer {
-	return certv1.Issuer{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: ns,
-		},
-		Spec: certv1.IssuerSpec{
-			IssuerConfig: certv1.IssuerConfig{
-				CA: &certv1.CAIssuer{
-					SecretName: LocalCACertSecretName,
-				},
-			},
-		},
-	}
+	certPEM := new(bytes.Buffer)
+	pem.Encode(certPEM, &pem.Block{
+		Type:  "CERTIFICATE",
+		Bytes: certBytes,
+	})
+
+	certPrivKeyPEM := new(bytes.Buffer)
+	pkcs8, _ := x509.MarshalPKCS8PrivateKey(certPrivKey)
+	pem.Encode(certPrivKeyPEM, &pem.Block{
+		Type:  "RSA PRIVATE KEY",
+		Bytes: pkcs8,
+	})
+
+	return certPEM, certPrivKeyPEM
 }